fixup to use floats instead of ints

+++ vector_add_kernel.cl
@@ -1,4 +1,4 @@
-__kernel void vector_add(__global const int *A, __global const int *B, __global int *C) {
+__kernel void vector_add(__global const float *A, __global const float *B, __global float *C) {
  
     // Get the index of the current element to be processed
     int i = get_global_id(0);
